solve: 바이러스
